update CreateUserRecord event

+++ assets/eip-5006/contracts/ERC5006.sol
@@ -14,9 +14,9 @@ contract ERC5006 is ERC1155, ERC1155Receiver, IERC5006 {
     mapping(uint256 => mapping(address => EnumerableSet.UintSet))
         private _userRecordIds;
     uint256 _curRecordId;
-    uint8 recordLimit;
+    uint256 recordLimit;
 
-    constructor(string memory uri_, uint8 recordLimit_) ERC1155(uri_) {
+    constructor(string memory uri_, uint256 recordLimit_) ERC1155(uri_) {
         recordLimit = recordLimit_;
     }
